+++ src/Commands.cpp
@@ -268,6 +268,7 @@ static SeqStrings gCommandNames =
     "CmdDeleteFileAndOpenNext\0"
     "CmdShowGeneratedHTML\0"
     "CmdDeleteCachedFiles\0"
+    "CmdToggleKeyboardLinkFollowing\0"
     "CmdNone\0"
     "\0";
 
@@ -529,6 +530,7 @@ static i32 gCommandIds[] = {
     CmdDeleteFileAndOpenNext,
     CmdShowGeneratedHTML,
     CmdDeleteCachedFiles,
+    CmdToggleKeyboardLinkFollowing,
     CmdNone,
 };
 
@@ -790,6 +792,7 @@ SeqStrings gCommandDescriptions =
     "Delete File And Open Next\0"
     "Show Generated HTML\0"
     "Delete Cached Files\0"
+    "Follow Link With Keyboard\0"
     "Do nothing\0"
     "\0";
 // clang-format on