+++ tests/all.ts
@@ -37,6 +37,7 @@ import { testit as issue5736 } from "./issue-5736.ts";
 import { testit as issue5751 } from "./issue-5751.ts";
 import { testit as issue5780 } from "./issue-5780.ts";
 import { testit as issue5529 } from "./issue-5529.ts";
+import { testit as issue2629 } from "./issue-2629.ts";
 import { testit as issue2799 } from "./issue-2799.ts";
 import { testit as findMatchSelect } from "./issue-find-match-select.ts";
 import { testit as issue2252 } from "./issue-2252.ts";
@@ -92,6 +93,7 @@ const tests: [string, () => void | Promise<void>][] = [
   ["issue-5751", issue5751],
   ["issue-5780", issue5780],
   ["issue-5529", issue5529],
+  ["issue-2629", issue2629],
   ["issue-2799", issue2799],
   ["issue-find-match-select", findMatchSelect],
   ["issue-2252", issue2252],